+++ core/src/main/java/feign/Util.java
@@ -139,7 +139,7 @@ public class Util {
    * Returns an unmodifiable collection which may be empty, but is never null.
    */
   public static <T> Collection<T> valuesOrEmpty(Map<String, Collection<T>> map, String key) {
-    return map.containsKey(key) ? map.get(key) : Collections.<T>emptyList();
+    return map.containsKey(key) && map.get(key) != null ? map.get(key) : Collections.<T>emptyList();
   }
 
   public static void ensureClosed(Closeable closeable) {